refactor: add specifications for bitmap-backed ID allocator and update dependencies
