+++ ostd/src/io/io_port/allocator.rs
@@ -3,6 +3,7 @@
 use vstd::prelude::*;
 use vstd::resource::set::{GhostSetAuth, GhostSubset};
 use vstd::tokens::InstanceId;
+use vstd_extra::external::{id_alloc_capacity, id_alloc_view};
 
 use core::ops::Range;
 
@@ -18,23 +19,12 @@ use crate::{
 
 verus! {
 
-/// Verus model for the external bitmap-backed ID allocator.
-#[verifier::external_type_specification]
-#[verifier::external_body]
-pub struct ExIdAlloc(IdAlloc);
-
 /// Opaque specification for the third-party one-time initialization primitive.
 #[verifier::external_type_specification]
 #[verifier::external_body]
 #[verifier::reject_recursive_types(T)]
 #[verifier::reject_recursive_types(R)]
 pub struct ExOnce<T, R>(spin::once::Once<T, R>);
-
-/// IDs currently allocated by an external `IdAlloc`.
-pub uninterp spec fn id_alloc_view(allocator: &IdAlloc) -> Set<usize>;
-
-/// Capacity configured for an external `IdAlloc`.
-pub uninterp spec fn id_alloc_capacity(allocator: &IdAlloc) -> usize;
 
 /// Identity assigned to the single global PIO allocator during trusted boot initialization.
 pub uninterp spec fn io_port_allocator_instance_id() -> InstanceId;
@@ -48,47 +38,6 @@ closed spec fn io_port_inner_inv_values(
     &&& allocated.subset_of(id_alloc_view(allocator))
     &&& id_alloc_capacity(allocator) == crate::arch::io::MAX_IO_PORT as usize
 }
-
-pub assume_specification[ IdAlloc::with_capacity ](capacity: usize) -> (allocator: IdAlloc)
-    ensures
-        id_alloc_capacity(&allocator) == capacity,
-        id_alloc_view(&allocator) == Set::<usize>::empty(),
-;
-
-pub assume_specification[ IdAlloc::is_allocated ](allocator: &IdAlloc, id: usize) -> (allocated:
-    bool)
-    ensures
-        id < id_alloc_capacity(allocator) ==> allocated == id_alloc_view(allocator).contains(id),
-;
-
-pub assume_specification[ IdAlloc::alloc_specific ](allocator: &mut IdAlloc, id: usize) -> (result:
-    Option<usize>)
-    requires
-        id < id_alloc_capacity(old(allocator)),
-    ensures
-        id_alloc_capacity(final(allocator)) == id_alloc_capacity(old(allocator)),
-        id_alloc_view(final(allocator)) == id_alloc_view(old(allocator)).insert(id),
-        id_alloc_view(old(allocator)).subset_of(id_alloc_view(final(allocator))),
-        id_alloc_view(old(allocator)).contains(id) ==> {
-            &&& result is None
-        },
-        !id_alloc_view(old(allocator)).contains(id) ==> {
-            &&& result == Some(id)
-        },
-    no_unwind
-;
-
-pub assume_specification[ IdAlloc::free_consecutive ](allocator: &mut IdAlloc, range: Range<usize>)
-    requires
-        range.end <= id_alloc_capacity(old(allocator)),
-    ensures
-        id_alloc_capacity(final(allocator)) == id_alloc_capacity(old(allocator)),
-        forall|id: usize| #[trigger]
-            id_alloc_view(final(allocator)).contains(id) <==> id_alloc_view(
-                old(allocator),
-            ).contains(id) && !(range.start <= id < range.end),
-    no_unwind
-;
 
 } // verus!
 /// Transparent facade used only to pass ghost frame facts to the third-party mutation.
